recolored and added the piechart to the peaktime
