+++ .cph/.A_Community_Education_Division.py_7e4043caa5af04bb25cd3ea562f303f3.prob
@@ -1,0 +1,1 @@
+{"name":"A. Community Education Division","group":"Codeforces - A2SV G5 - Contest #6 | Ethiopian Christmas Special!","url":"https://codeforces.com/gym/496610/problem/A","interactive":false,"memoryLimit":256,"timeLimit":1000,"tests":[{"id":1704526626944,"input":"7\n-789\n1299\n1300\n1399\n1400\n1679\n2300","output":"Division 4\nDivision 4\nDivision 4\nDivision 4\nDivision 3\nDivision 2\nDivision 1"}],"testType":"single","input":{"type":"stdin"},"output":{"type":"stdout"},"languages":{"java":{"mainClass":"Main","taskClass":"ACommunityEducationDivision"}},"batch":{"id":"fdb911e2-832f-41d9-a115-1f3d0309c228","size":5},"srcPath":"c:\\6kiloCafeProject\\A_Community_Education_Division.py"}
