+++ .cph/.B_Encrypted_Postcards.py_12dfd2a89bb80446f7139166157d3a92.prob
@@ -1,0 +1,1 @@
+{"name":"B. Encrypted Postcards","group":"Codeforces - A2SV G5 - Contest #6 | Ethiopian Christmas Special!","url":"https://codeforces.com/gym/496610/problem/B","interactive":false,"memoryLimit":256,"timeLimit":1000,"tests":[{"input":"3\n8\nhbhihbhi\n5\nozxco\n20\nmmeellkkaammbbeeaall\n","output":"hi\no\nmelkambeal\n","id":1704522029931},{"id":1704522029944,"input":"2\n50\nehjfbsssjkfhenopuoofjkjsjheenkdakuvvuazzaniiqwewdn\n30\naiadlakdezzierrefklkkeskseehih","output":"enkuan\nadereseh"}],"testType":"single","input":{"type":"stdin"},"output":{"type":"stdout"},"languages":{"java":{"mainClass":"Main","taskClass":"BEncryptedPostcards"}},"batch":{"id":"0cce5b02-c2a5-41d5-ace2-156b1db9bcb0","size":1},"srcPath":"c:\\6kiloCafeProject\\B_Encrypted_Postcards.py"}
