+++ .cph/.C_Gena.py_6b15562f51d9995e99fd590b4c84ba62.prob
@@ -1,0 +1,1 @@
+{"name":"C. Gena","group":"Codeforces - A2SV G5 - Contest #6 | Ethiopian Christmas Special!","url":"https://codeforces.com/gym/496610/problem/C","interactive":false,"memoryLimit":256,"timeLimit":1000,"tests":[{"input":"4\n1 4 6 2\n5\n5 1 5 7 9\n","output":"3\n","id":1704522672552},{"input":"4\n1 2 3 4\n4\n10 11 12 13\n","output":"0\n","id":1704522672540},{"id":1704522672597,"input":"5\n1 1 1 1 1\n3\n1 2 3","output":"2"}],"testType":"single","input":{"type":"stdin"},"output":{"type":"stdout"},"languages":{"java":{"mainClass":"Main","taskClass":"CGena"}},"batch":{"id":"82b227e9-fd25-4cd3-a128-8e11086fa474","size":1},"srcPath":"c:\\6kiloCafeProject\\C_Gena.py"}
