+++ .cph/.D_Metsehafe_and_Enemy_Aliens.py_3df2c3047e1e8fab43f50d65172ea405.prob
@@ -1,0 +1,1 @@
+{"name":"D. Metsehafe and Enemy Aliens","group":"Codeforces - A2SV G5 - Contest #6 | Ethiopian Christmas Special!","url":"https://codeforces.com/gym/496610/problem/D","interactive":false,"memoryLimit":256,"timeLimit":1000,"tests":[{"id":1704527066242,"input":"4\n4\n1 3 1 1\n4\n1 2 1 1\n6\n3 2 1 5 2 4\n2\n1 6","output":"4\n4\n11\n5"},{"id":1704527716073,"input":"1\n4\n1 1 1 5","output":"5"}],"testType":"single","input":{"type":"stdin"},"output":{"type":"stdout"},"languages":{"java":{"mainClass":"Main","taskClass":"DMetsehafeAndEnemyAliens"}},"batch":{"id":"1f3c7bf3-e8ba-46e3-8028-b4e7c4835a55","size":1},"srcPath":"c:\\6kiloCafeProject\\D_Metsehafe_and_Enemy_Aliens.py"}
