+++ .cph/.E_m_Lucky_Numbers.py_ec2b320cf0235f30303b1c9e7edc60e5.prob
@@ -1,0 +1,1 @@
+{"name":"E. m-Lucky Numbers","group":"Codeforces - A2SV G5 - Contest #6 | Ethiopian Christmas Special!","url":"https://codeforces.com/gym/496610/problem/E","interactive":false,"memoryLimit":256,"timeLimit":1000,"tests":[{"id":1704525170566,"input":"3\n5\n1 2 3 4 5\n5\n4 4 4 4 2\n6\n1 3 1 5 3 1","output":"-1 -1 3 2 1\n-1 4 4 4 2\n-1 -1 1 1 1 1"}],"testType":"single","input":{"type":"stdin"},"output":{"type":"stdout"},"languages":{"java":{"mainClass":"Main","taskClass":"EMLuckyNumbers"}},"batch":{"id":"00052886-2e7c-4fda-b48d-4ef424a258fa","size":1},"srcPath":"c:\\6kiloCafeProject\\E_m_Lucky_Numbers.py"}
